Moved the date validation function to utils.
